filed datamodel to migrate and moved anomalydetector to anomaly_detection
