+++ airflow-core/docs/img/airflow_erd.sha256
@@ -1,1 +1,1 @@
-6359bf94b1b5e15156a608179073ce97e8c9424c49f53433d57cbd2446eed62a
+558818c8ad4e3e0b477bed98741f91651ec0b2f7e48da390bedac8be7c184fe3
