+++ airflow-core/docs/migrations-ref.rst
@@ -39,7 +39,7 @@ Here's the list of all the Database Migrations that are executed via when you ru
 +-------------------------+------------------+-------------------+--------------------------------------------------------------+
 | Revision ID             | Revises ID       | Airflow Version   | Description                                                  |
 +=========================+==================+===================+==============================================================+
-| ``7e8e6d53a4b2`` (head) | ``53ff648b8a26`` | ``3.2.0``         | Add dag_run_anomaly table.                                   |
+| ``438ed3578bfb`` (head) | ``53ff648b8a26`` | ``3.2.0``         | Add task instance anomaly table.                             |
 +-------------------------+------------------+-------------------+--------------------------------------------------------------+
 | ``53ff648b8a26``        | ``a5a3e5eb9b8d`` | ``3.2.0``         | Add revoked_token table.                                     |
 +-------------------------+------------------+-------------------+--------------------------------------------------------------+
