+++ airflow-core/src/airflow/anomaly_detection/anomalydetector.py
@@ -152,7 +152,7 @@ class AnomalyDetector:
 #     def callback(context):
 #         if existing_callback:
 #             existing_callback(context)
-
+#
 #         detector.trigger_anomaly_detection(context)
-
+#
 #     task.on_success_callback = callback
