+++ airflow-core/src/airflow/models/__init__.py
@@ -66,7 +66,6 @@ def import_all_models():
     import airflow.models.dag_version
     import airflow.models.dagbag
     import airflow.models.dagbundle
-    import airflow.models.dagrun_anomaly
     import airflow.models.dagwarning
     import airflow.models.deadline_alert
     import airflow.models.errors
